improved error handler

+++ cogs/scraper.py
@@ -60,7 +60,7 @@ class Scraper(commands.Cog):
             details = i.find("div", {"class": "aditem-details"})
             ad.price = details.find("strong").text
             radius = details.contents[8].split()[-2:]
-            if int(radius[0]) > config["radius"]:
+            if float(radius[0]) > config["radius"]:
                 continue
             ad.location = f"{details.contents[4].split()[-1]} {details.contents[6].split()[-1]} ({' '.join(radius)})"
             ad.description = i.find(
@@ -102,5 +102,6 @@ class Scraper(commands.Cog):
         channel = self.bot.get_channel(config.LOG_CHANNEL_ID)
         await channel.send(embed=simple_embed(self.bot.user, "scraper error", color=discord.Color.orange()))
         await on_command_error(self.bot.get_channel(config.LOG_CHANNEL_ID), error)
+        
 def setup(bot):
     bot.add_cog(Scraper(bot))
